Add LIDAR and LidarSpin classes Add BoilerTracker class that rotates the Camera/LIDAR turntable to track the boiler vision target.
If the vision target is tracked, the LIDAR distance is the radius of a circle centered on the boiler.  The turntable angle can be added to the IMU yaw angle and converted to a filed relative angle to the boiler.

This polar coordinate relative to the boiler is then converted to cartesian field coordinates and published as Robot.currentRobotFieldPosition