Now using a customized Encoder class that checks the scan results every time the encoder count changes, (about 60 times as often as vision updates).
We start scanning less often, (when vision frames arrive), and most likely stop the scan in response to an interrupt that indicates the scan has passed the desired angle.

This will leave the turntable right in the center of the dead zone so the next vision update will not start the scan again unless the robot has turned more than the deadzone angle since the last vision update.