RobotPosition is now stored and used internally in polar coordinates.
Errors introduced by the IMU over the course of the match do not affect the radius measured by the LIDAR, so the boiler tracker can be used reliably throughout the game as long as you stick to driving along radial arcs.  The feeder that cant be see from 2 of 3 driver stations and two of the three ropes are easily approached using an arc to navigate.